+++ packages/frontend/src/components/common/YearInput.tsx
@@ -14,7 +14,7 @@ const YearInput: FC<Props> = ({ value, onChange, ...props }) => {
 
     return (
         <YearPickerInput
-            sx={{ width: '100%' }}
+            w="100%"
             size="xs"
             minDate={moment().year(1000).toDate()}
             maxDate={moment().year(9999).toDate()}