+++ packages/frontend/src/providers/ThirdPartyServicesProvider.tsx
@@ -47,7 +47,9 @@ const ThirdPartyServicesEnabledProvider: FC = ({ children }) => {
                     api_host: health.data?.posthog.apiHost,
                     bootstrap: {
                         featureFlags: {
+                            'chart-version-history': false,
                             'lightdash-team-flair': false,
+                            'extended-usage-analytics': false,
                         },
                     },
                 }}